requirements.txt
PyMuPDF==1.24.13
llama-cpp-python==0.3.1
huggingface-hub==0.26.2